% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partitionR.R
\name{partitionR}
\alias{partitionR}
\title{Partitioning community CV}
\usage{
partitionR(z, n = 1)
}
\arguments{
\item{z}{A \code{matrix} containing repeated measurements of species abundances.
The \code{matrix} must contain numerical values only, with years in rows and species in
columns. Remove any extra column.}

\item{n}{Only species appearing more than \code{n} years (\code{integer}, defaults to 1) are used in the calculations.}
}
\value{
Returns an object of class \code{'comstab'}. It contains the mean coefficient of variation of species (\eqn{\overline{CV}}),
the dominance (\eqn{ \Delta}), asynchrony (\eqn{ \Psi}) and averaging (\eqn{ \omega}) effects.
}
\description{
\code{PartitionR()} is a function used to partition the temporal coefficient
of variation of a community into the mean coefficient of variation of species
and three stabilizing components: the dominance, asynchrony and averaging effects
(see Details).
}
\details{
The analytic framework is described in details in Segrestin \emph{et al.} (in prep).
In short, the partitioning relies on the following equation: \deqn{CV=\overline{CV} \Delta \Psi \omega}
where \eqn{CV} is the community coefficient of variation (reciprocal of community stability),
\eqn{\overline{CV}} is the mean species coefficient of variation, \eqn{ \Delta} is the dominance effect,
\eqn{ \Psi} is the asynchrony effect, and \eqn{ \omega} is the averaging effect.
}
\examples{
require(stats)

# Simulates a custom community time series using 'comTS()':
z <- comTS(nsp = 10, ny = 30, even = 0.6, mvs = 1.5, sync = "0")

# Runs the partitioning of the community coefficient of variation:
partitionR(z)

}
\references{
Segrestin \emph{et al.} (in prep) A unified framework for partitioning the drivers of stability of ecological communities
}
\author{
Jules Segrestin, \email{jsegrestin@gmail.com}
}
