% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.TPL}
\alias{plot.TPL}
\title{Plotting a \code{TPL} object}
\usage{
\method{plot}{TPL}(x, \dots, xlab = "Mean", ylab = "Variance", log = "xy")
}
\arguments{
\item{x}{object of class \code{'comstab'}.}

\item{...}{other parameters to be passed through to plotting functions.}

\item{xlab}{a label for the x axis, defaults to 'Mean".}

\item{ylab}{a label for the y axis, defaults to 'Variance'.}

\item{log}{Defaults to 'xy' to show the log-log relationship.}
}
\value{
No return value, graphical function.
}
\description{
Plotting method for object inheriting from class "TPL".
}
\examples{
require(graphics)

# Simulates a custom community time series using 'comTS()':
z <- comTS(nsp = 10, ny = 30, even = 0.6, mvs = 1.5, sync = "0")

# Computes Taylor's power law:
x <- TPL(z)

# Plots the result
plot(x)

}
\author{
Jules Segrestin, \email{jsegrestin@gmail.com}
}
