% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TPL.R
\name{TPL}
\alias{TPL}
\title{Computing Taylor's power law}
\usage{
TPL(z, ny = 1, method = c("OLS", "SMA"))
}
\arguments{
\item{z}{A \code{matrix} containing repeated measurements of species abundances.
The \code{matrix} must contain numerical values only, with years in rows and species in
columns. Remove any extra column.}

\item{ny}{Only species appearing more than \code{ny} years (\code{integer}, defaults to 1) are used in the calculations.}

\item{method}{Regression method used to estimate the scaling coefficient \code{(factor)}:
\itemize{
\item \code{OLS} (defaults) : Ordinary least square
\item \code{SMA}: Standardized major axis, sometimes called reduced major axis
}}
}
\value{
An object of class \code{'TPL'} is a list containing the following components:
\itemize{
\item \code{'data'} a \code{data.frame} containing species temporal means and variance
\item \code{'test'} the result of the Pearson's correlation test between log10-transformed species means and variances (see \code{cor.test} in the \code{stats} package)
\item \code{a} the intercept of the power law
\item \code{b} the scaling coefficient of the power law
}
}
\description{
\code{TPL()} is a function used to compute the power function between
species temporal means and variances.
}
\examples{
require(smatr)

# Simulates a custom community time series using 'comTS()':
z <- comTS(nsp = 10, ny = 30, even = 0.6, mvs = 1.5, sync = "0")

# Computes Taylor's power law:
TPL(z)

}
\author{
Jules Segrestin, \email{jsegrestin@gmail.com}
}
